+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The EnhancedUnrealEngineDocumentation plugin adds a convenient way to get improv
 
 ### Added
 
+- [ISSUE-9](https://github.com/JetBrains/EnhancedUnrealEngineDocumentation/issues/9) Support for meta specifiers in UE reflections
+
 ### Changed
 
 ### Deprecated